+++ src/parsetree/expressions/expressions.js
@@ -521,8 +521,11 @@ export class InExpression extends Expression {
     static parse(parser, root) {
         if (!parser.matchToken("in")) return;
         var target = parser.requireElement("unaryExpression");
-        var inExpression = new InExpression(root, target);
-        return parser.parseElement("indirectExpression", inExpression);
+        var result = new InExpression(root, target);
+        if (parser.matchToken("where")) {
+            result = new WhereExpression(result, CollectionExpression.parseOperand(parser));
+        }
+        return parser.parseElement("indirectExpression", result);
     }
 
     resolve(context, { root: rootVal, target }) {
@@ -530,7 +533,7 @@ export class InExpression extends Expression {
         var returnArr = [];
         if (rootVal.css) {
             context.meta.runtime.implicitLoop(target, function (targetElt) {
-                var results = targetElt.querySelectorAll(rootVal.css);
+                var results = context.meta.runtime.resolveQuery(targetElt, rootVal.css);
                 for (var i = 0; i < results.length; i++) {
                     returnArr.push(results[i]);
                 }